+++ src/Classifiers/SoftmaxClassifier.php
@@ -86,11 +86,12 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
     protected $costFn;
 
     /**
-     * The unique class labels.
+     * The number of epochs without improvement in the training loss to wait
+     * before considering an early stop.
      *
-     * @var array|null
+     * @var int
      */
-    protected $classes;
+    protected $window;
 
     /**
      * The underlying neural network instance.
@@ -98,6 +99,13 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
      * @var \Rubix\ML\NeuralNet\FeedForward|null
      */
     protected $network;
+
+    /**
+     * The unique class labels.
+     *
+     * @var array|null
+     */
+    protected $classes;
 
     /**
      * The average cost of a training sample at each epoch.
@@ -115,6 +123,7 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
      * @param int $epochs
      * @param float $minChange
      * @param \Rubix\ML\NeuralNet\CostFunctions\ClassificationLoss|null $costFn
+     * @param int $window
      * @throws \InvalidArgumentException
      */
     public function __construct(
@@ -123,7 +132,8 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
         float $alpha = 1e-4,
         int $epochs = 1000,
         float $minChange = 1e-4,
-        ?ClassificationLoss $costFn = null
+        ?ClassificationLoss $costFn = null,
+        int $window = 5
     ) {
         if ($batchSize < 1) {
             throw new InvalidArgumentException('Cannot have less than 1 sample'
@@ -145,12 +155,18 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
                 . " than 0, $minChange given.");
         }
 
+        if ($window < 1) {
+            throw new InvalidArgumentException('Window must be at least 1'
+                . " epoch, $window given.");
+        }
+
         $this->batchSize = $batchSize;
         $this->optimizer = $optimizer ?? new Adam();
         $this->alpha = $alpha;
         $this->epochs = $epochs;
         $this->minChange = $minChange;
         $this->costFn = $costFn ?? new CrossEntropy();
+        $this->window = $window;
     }
 
     /**
@@ -206,6 +222,8 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
     }
 
     /**
+     * Train the learner with a dataset.
+     *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException
      */
@@ -231,8 +249,7 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
     }
 
     /**
-     * Perform mini-batch gradient descent with given optimizer over the training
-     * set and update the input weights accordingly.
+     * Perform a partial train on the learner.
      *
      * @param \Rubix\ML\Datasets\Dataset $dataset
      * @throws \InvalidArgumentException
@@ -263,7 +280,8 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
             ]));
         }
 
-        $prevLoss = INF;
+        $prevLoss = $bestLoss = INF;
+        $delta = 0;
 
         for ($epoch = 1; $epoch <= $this->epochs; $epoch++) {
             $batches = $dataset->randomize()->batch($this->batchSize);
@@ -287,6 +305,18 @@ class SoftmaxClassifier implements Estimator, Online, Probabilistic, Verbose, Pe
             }
 
             if (abs($prevLoss - $loss) < $this->minChange) {
+                break 1;
+            }
+
+            if ($loss < $bestLoss) {
+                $bestLoss = $loss;
+                
+                $delta = 0;
+            } else {
+                $delta++;
+            }
+
+            if ($delta >= $this->window) {
                 break 1;
             }